Adding base files (#1)
* Add stock data

This includes a subset of the truth source dataset. This is done for two reasons: 1. The truth source is too large to upload. 2. Running our script on this smaller dataset will be faster.

* Add initial VBA script

This is my initial rough draft on how we may want to approach the problem. There are no comments or ideas outlined. Those are to be added in the next commit.

* Add meta process file

This file indicates all the sources and collaboration when writing the vba-challenge.vbs script. Additionally, this file includes additional planning and thoughts not included in comments in the script.