Update vba-challenge.vbs
Tested out basic subroutines to identify the last row in each the dataset (to loop through) and to loop through each sheet in the workbook. Integrated these into the main StockSummary() subroutine. Still running into an issue where one of the calculations creates a value that is too large to be stored in a double.